Permission

+++ resources/views/dashboard/partial/exchange_rates.blade.php
@@ -4,9 +4,11 @@
     {{-- Table Title --}}
     <x-slot:title>
         <h5 class="card-header">@lang('index.rates.exchange_rates')</h5>
-        <button type="button" data-bs-toggle="modal" data-bs-target="#ModalAddRate" class="btn rounded-pill btn-icon btn-label-primary">
-            <span class="tf-icons bx bx-plus"></span>
-        </button>
+        @can('exchange-rate-create')
+            <button type="button" data-bs-toggle="modal" data-bs-target="#ModalAddRate" class="btn rounded-pill btn-icon btn-label-primary">
+                <span class="tf-icons bx bx-plus"></span>
+            </button>
+        @endcan
     </x-slot:title>
 
     {{-- Table Header --}}
@@ -37,19 +39,23 @@
                 </td>
                 <td>
                     <x-dropdown-table>
-                        <button class="dropdown-item" onclick="edit_rate({{ $rate->id }})" type="button" data-bs-toggle="modal" data-bs-target="#ModalEditRate" href="javascript:void(0);"><i class="bx bx-edit-alt me-1"></i>
-                            @lang('general.edit')
-                        </button>
-                        <button class="dropdown-item" onclick="delete_rate({{ $rate->id }})" type="button"  data-bs-toggle="modal"  data-bs-target="#ModalDeleteRate" href="javascript:void(0);"><i class="bx bx-trash me-1"></i>
-                            @lang('general.delete')
-                        </button>
-                        <button class="dropdown-item" type="button" onclick="active_rate({{ $rate->id }})" href="javascript:void(0);"><i class="bx bx-edit-alt me-1"></i>
-                           @if($rate->is_active)
-                                @lang('general.deactivation')
-                            @else
-                                @lang('general.active')
-                            @endif
-                        </button>
+                        @can('exchange-rate-edit')
+                            <button class="dropdown-item" onclick="edit_rate({{ $rate->id }})" type="button" data-bs-toggle="modal" data-bs-target="#ModalEditRate" href="javascript:void(0);"><i class="bx bx-edit-alt me-1"></i>
+                                @lang('general.edit')
+                            </button>
+                        @endcan
+                        @can('exchange-rate-delete')
+                                <button class="dropdown-item" onclick="delete_rate({{ $rate->id }})" type="button"  data-bs-toggle="modal"  data-bs-target="#ModalDeleteRate" href="javascript:void(0);"><i class="bx bx-trash me-1"></i>
+                                    @lang('general.delete')
+                                </button>
+                                <button class="dropdown-item" type="button" onclick="active_rate({{ $rate->id }})" href="javascript:void(0);"><i class="bx bx-edit-alt me-1"></i>
+                                    @if($rate->is_active)
+                                        @lang('general.deactivation')
+                                    @else
+                                        @lang('general.active')
+                                    @endif
+                                </button>
+                        @endcan
                     </x-dropdown-table>
                 </td>
             </tr>
